Extract FormFieldButton as a component
