+++ src/components/person/PersonSelectDialog.tsx
@@ -1,18 +1,10 @@
 import { LoadingButton } from '@mui/lab'
-import {
-  Box,
-  Button,
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogTitle,
-  Typography,
-} from '@mui/material'
-import { makeStyles } from '@mui/styles'
+import { Box, Dialog, DialogActions, DialogContent, DialogTitle } from '@mui/material'
 import { translateError } from 'common/form/validation'
 import useConfirm from 'common/hooks/confirm'
 import theme from 'common/theme'
 import CloseModalButton from 'components/common/CloseModalButton'
+import FormFieldButton from 'components/common/FormFieldButton'
 import { PersonResponse } from 'gql/person'
 import { useTranslation } from 'next-i18next'
 import React, { useState } from 'react'
@@ -25,37 +17,9 @@ type Props = {
   error?: string
 }
 
-const useStyles = makeStyles({
-  imitateInputBox: {
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    border: `1px solid rgba(0, 0, 0, 0.23)`,
-    borderRadius: '3px',
-    padding: '8.5px 14px',
-    cursor: 'pointer',
-  },
-  errorInputBox: {
-    borderColor: '#d32f2f',
-    color: '#d32f2f',
-  },
-  errorText: {
-    color: '#d32f2f',
-    fontWeight: 400,
-    fontSize: '0.75rem',
-    lineHeight: 1.66,
-    letterSpacing: '0.03333em',
-    textAlign: 'left',
-    marginTop: '4px',
-    marginRight: '14px',
-    marginBottom: 0,
-    marginLeft: '14px',
-  },
-})
 function PersonSelectDialog({ onConfirm: confirmCallback, onClose: closeCallback, error }: Props) {
   const [person, setPerson] = useState<PersonResponse | null>(null)
-  const { t } = useTranslation('person')
-  const classes = useStyles()
+  const { t } = useTranslation()
   const { open, confirmHandler, closeHandler, openHandler, loading } = useConfirm({
     onConfirm: async () => {
       confirmCallback ? confirmCallback(person) : null
@@ -67,21 +31,13 @@ function PersonSelectDialog({ onConfirm: confirmCallback, onClose: closeCallback
 
   return (
     <>
-      <Box
+      <FormFieldButton
         onClick={openHandler}
-        className={
-          error ? classes.imitateInputBox + ' ' + classes.errorInputBox : classes.imitateInputBox
-        }>
-        <Typography>
-          {person
-            ? `${person.firstName} ${person.lastName} (${person.id})`
-            : t('person:selectDialog.notSelected')}
-        </Typography>
-        <Button sx={{ padding: 0 }} onClick={openHandler}>
-          {t('person:selectDialog.select')}
-        </Button>
-      </Box>
-      {error ? <p className={classes.errorText}>{translateError(error, t)}</p> : null}
+        placeholder={t('person:selectDialog.notSelected')}
+        value={person ? `${person.firstName} ${person.lastName} (${person.id})` : undefined}
+        button={{ label: t('person:selectDialog.select') }}
+        error={error ? translateError(error, t) : undefined}
+      />
       <Dialog fullWidth open={open} onClose={closeHandler}>
         <DialogTitle>{t('person:selectDialog.personSelect')}</DialogTitle>
         <DialogContent>
